Renamed result_screen to barcode_result_screen.
