+++ lib/screens/barcode_result_screen.dart
